fix(app): restore the inline media selections, reverting the shared fragment
The shared `@inline` fragment silently dropped media everywhere it was read
structurally. Relay stores `@inline` data under `__fragments` rather than flattening
it onto the part:

    {"__typename":"FileContentPart",
     "__fragments":{"mediaContentPartFragment":{"file":{...,"variable":"files"}}}}

so `asFilePart` and friends found no `file` key and returned nothing. The prompt page
rendered a media-only message as an empty card, and — worse — loading a prompt into
the playground dropped its media, which means a load-then-save erased it. Exactly the
data loss the selection guard exists to prevent.

Nothing caught it. The selection was present, so the guard passed; the types compiled,
because Relay's generated union ends in a `"%other"` arm; and the tests fed plain
objects rather than Relay data, so they exercised the one shape that still worked. I
had checked the generated reader contained the media fields and concluded the data was
flattened — the reader says what is selected, not where it lands.

The five documents ask for the media inline again. `readMessageMedia` groups the parts
structurally, which is one shape for data from Relay and data built by hand. The
guard is back to requiring the selection in the document, and now says why accepting a
spread is not an option: the duplication is the cheaper problem, and the guard is what
keeps the copies honest.

Costs 41 added lines across four upstream-owned files, which is the right trade.
